fix(web): satisfy stats page lint gate
Extract Stats option constants into a helper module for fast refresh compliance.
